+++ lib/mesa/src/gallium/drivers/i915/i915_resource_texture.c
@@ -155,8 +155,8 @@ i915_texture_set_image_offset(struct i915_texture *tex, unsigned level,
    tex->image_offset[level][img].nblocksy = nblocksy;
 
 #if DEBUG_TEXTURES
-   debug_printf("%s: %p level %u, img %u (%u, %u)\n", __func__, tex, level,
-                img, x, y);
+   debug_printf("%s: %p level %u, img %u (%u, %u)\n", __func__, tex, level, img,
+                x, y);
 #endif
 }
 
@@ -208,12 +208,10 @@ i9x5_scanout_layout(struct i915_texture *tex)
    i915_texture_set_level_info(tex, 0, 1);
    i915_texture_set_image_offset(tex, 0, 0, 0, 0);
 
-
 #if DEBUG_TEXTURES
-   debug_printf("%s size: %d,%d,%d offset %d,%d (0x%x)\n", __func__,
-                pt->width0, pt->height0, util_format_get_blocksize(pt->format),
-                tex->stride, tex->total_nblocksy,
-                tex->stride * tex->total_nblocksy);
+   debug_printf("%s size: %d,%d,%d offset %d,%d (0x%x)\n", __func__, pt->width0,
+                pt->height0, util_format_get_blocksize(pt->format), tex->stride,
+                tex->total_nblocksy, tex->stride * tex->total_nblocksy);
 #endif
 
    return true;
@@ -242,10 +240,9 @@ i9x5_display_target_layout(struct i915_texture *tex)
    tex->tiling = I915_TILE_X;
 
 #if DEBUG_TEXTURES
-   debug_printf("%s size: %d,%d,%d offset %d,%d (0x%x)\n", __func__,
-                pt->width0, pt->height0, util_format_get_blocksize(pt->format),
-                tex->stride, tex->total_nblocksy,
-                tex->stride * tex->total_nblocksy);
+   debug_printf("%s size: %d,%d,%d offset %d,%d (0x%x)\n", __func__, pt->width0,
+                pt->height0, util_format_get_blocksize(pt->format), tex->stride,
+                tex->total_nblocksy, tex->stride * tex->total_nblocksy);
 #endif
 
    return true;
@@ -893,7 +890,7 @@ i915_texture_transfer_unmap(struct pipe_context *pipe,
 void
 i915_texture_subdata(struct pipe_context *pipe, struct pipe_resource *resource,
                      unsigned level, unsigned usage, const struct pipe_box *box,
-                     const void *data, unsigned stride, unsigned layer_stride)
+                     const void *data, unsigned stride, uintptr_t layer_stride)
 {
    /* i915's cube and 3D maps are not laid out such that one could use a
     * layer_stride to get from one layer to the next, so we have to walk the
